Add `Transaction::set_account` method
